+++ backend/src/main/java/hanglog/community/service/CommunityService.java
@@ -7,9 +7,9 @@ import static hanglog.trip.domain.type.PublishedStatusType.PUBLISHED;
 import hanglog.auth.domain.Accessor;
 import hanglog.city.domain.City;
 import hanglog.city.domain.repository.CityRepository;
+import hanglog.city.dto.CityElements;
 import hanglog.community.domain.recommendstrategy.RecommendStrategies;
 import hanglog.community.domain.recommendstrategy.RecommendStrategy;
-import hanglog.community.dto.CityElements;
 import hanglog.community.dto.response.CommunityTripListResponse;
 import hanglog.community.dto.response.CommunityTripResponse;
 import hanglog.community.dto.response.RecommendTripListResponse;
