+++ backend/src/main/java/hanglog/trip/domain/repository/TripCityRepository.java
@@ -1,6 +1,6 @@
 package hanglog.trip.domain.repository;
 
-import hanglog.community.dto.CityElement;
+import hanglog.city.dto.CityElement;
 import hanglog.trip.domain.TripCity;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,8 +9,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 public interface TripCityRepository extends JpaRepository<TripCity, Long> {
-
-    List<TripCity> findByTripId(Long tripId);
 
     @Modifying
     @Query("""
@@ -21,7 +19,7 @@ public interface TripCityRepository extends JpaRepository<TripCity, Long> {
     void deleteAllByTripId(@Param("tripId") final Long tripId);
 
     @Query("""
-            SELECT new hanglog.community.dto.CityElement (tc.trip.id, tc.city) 
+            SELECT new hanglog.city.dto.CityElement (tc.trip.id, tc.city) 
             FROM TripCity tc 
             WHERE tc.trip.id IN :tripIds
             """)
